Add support git-config like values
Basically, makes it possible to have lines like

         [thing "section"]
                  key = value1
                  key = value2

That evaluates to thing.section.key = [value1, value2].

As a result, this changes the result format to being more like
url.Values or what have you where it's a map of strings to slices of
strings. So, you can define multiple values for something.

Fairly useful, but weird.

Change-Id: I0feed4827069bbc1a878cbdae8560c51b38189c7