Add comments for most exported names
Covers exported functions, types, variables, and constants.

This has one code change, in that it changes ReadINI to use
a bytes.Reader instead of a bytes.Buffer. This is primarily to avoid
giving the decoder only a reader type without also providing a writer,
but also because it just makes more sense.

Given that ReadINI is a pre-rewrite function kept around for
compatibility, it's not the most urgent function present. In spite of
that, test readers now include a bytes.Reader in addition to
bytes.Buffer, strings.Reader, and so on.

Change-Id: I6cc240f4ec1a9e57801d05d5f1eca8e65dad1b4b